11 side navbar (#35)
* beginning SideNavBar component

created the SideNavBar component and began styling. all of the buttons have been created

* continued work on SideNavBar

added the bread & roses logo, adjusted some spacing, and began work on the footer at the bottom.

* finished component

* working on pr comments

began working with the next.js router

* more modular styling

made it so that certain buttons show up depending on if the user's role is admin or volunteer. also cleaned up the styling so that it looks cleaner and fixed up some spacing

* made new pages and rerouted buttons

* finished buttons!

* small changes

---------

Co-authored-by: vle04 <ngocvinale@gmail.com>
Co-authored-by: wkim10 <wkim10@tufts.edu>