69 frontend check incheck out flow (#83)
* updated SideNavBar

navigation bar now has tab for check-in/out

* updated check-in/out form

styled and formatted check in/check out form to match the figma. styling of buttons changes when clicked and only one button is able to be active at a time

* update CheckInOut.tsx

* styling confirm shift check-in stage

now replying on a useState to track what stage of the check-in flow the user is on. began styling the second stage of this flow (confirming shift check-in)

* update to shift signup modal

* finished shifts page

* removed unused imports

* created confirmation module for shift sign in/check out

* update

* autocomplete email

* only one shift can be selected at a time

* remove comment

* update

* small styling changes

* remove name field from flow

* reset selected shift time

---------

Co-authored-by: wkim10 <wkim10@tufts.edu>